+++ conf/broker.conf
@@ -392,7 +392,7 @@ preciseDispatcherFlowControl=false
 
 # Class name of the factory that implements the DispatcherProvider.
 # We can customize the Dispatcher through the implementation class.
-dispatcherProviderClassName=org.apache.pulsar.broker.service.DispatcherFactory.DefaultDispatcherProvider
+dispatcherProviderClassName=org.apache.pulsar.broker.service.DefaultDispatcherProvider
 
 # Max number of concurrent lookup request broker allows to throttle heavy incoming lookup traffic
 maxConcurrentLookupRequest=50000
